BackEnd/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
motor>=3.6.0
pymongo[srv]
pydantic[email]==2.6.3
pydantic-settings==2.2.1
python-dotenv==1.0.1
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
python-multipart
certifi==2024.8.30

